Add c integrations for parquet reader writer
