+++ nautilus_core/persistence/src/lib.rs
@@ -13,5 +13,4 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod c_api;
 pub mod parquet;
