+++ nautilus_trader/core/rust/persistence.pxd
